[Sparc] Add support for flat register window mode (-mflat)
With flat register mode the compiler assumes that there is only one register window. Instead
of using `save`/`restore` instructions when entering/leaving a function, the registers that
are used by the function are explictly saved to the stack at entry and restored at exit. This
adds extra work to each function, but will make the execution time more predictable as there
is no risk of a register window overflow or underflow.

Implementing this requires writing completely new functions for emitting function prologues
and epilogues. Instead of simply using a `save` instruction, each of the `%i` and `%l` registers
that are used by the function needs to be written to the stack, if possible with a `std` instructi
on for performance.

The dwarf call frame information provided in the prologue is relatively simple for the standard
Sparc register window model. It only requires a single `.cfi_window_save` instruction to specify
that the register window has shifted. For the flat register window model one instead needs to
specify for each register that is saved to the stack exactly where it is located. This information
is then used by both debuggers and the exception handling in C++.

With flat register window mode there is no longer any need for the renaming of registers that is
necessary when switching register window. This means that all code that expects registers to have
changed names will have to be rewritten to be a aware that this does not happen in flat register
window mode.