[Sparc] Use annulling branches if unable to fill delay slot
Branch always instructions can be directly replaced with a branch with
an annulled delay slot. The delay slot will never be executed and can
be ignored.

For other annulled branch instructions the delay slot is only executed
when the branch is taken. This means that the target instruction of the
branch can be placed in the delay slot and the branch can then skip the
original target instruction and take the next one instead.

To be able to skip the first instruction we need to split the block into
two, with the first block only including the first instruction.

If the target block is the same block that we are iterating over we will
run into problems with the iterator. For this reason we defer the split
until after all blocks in the function has been passed through.